Week1
This was the end result of our first week of work on the project. We used numerical matrix methods to solve the Schrodinger equation in infinite square wells and simple harmonic oscillator potentials.. The file is designed to be run on Matlab using the 'publish' to generate a HTML that explains the code.